INFUND-3719 Updates for review comments.
Former-commit-id: 3e120d5f14ca61fff8b2ae7df01f98002c663725

+++ ifs-web-service/ifs-assessment-service/src/main/java/com/worth/ifs/assessment/model/AssessorDashboardModelPopulator.java
@@ -49,8 +49,8 @@ public class AssessorDashboardModelPopulator {
                      new AssessorDashboardActiveCompetitionViewModel(
                             cpr.getCompetitionId(),
                             cpr.getCompetitionName(),
-                            cpr.getSubmittedAssessments().intValue(),
-                            cpr.getTotalAssessments().intValue(),
+                            cpr.getSubmittedAssessments(),
+                            cpr.getTotalAssessments(),
                             cpr.getAssessorDeadlineDate().toLocalDate(),
                             cpr.getAssessmentDaysLeft(),
                             cpr.getAssessmentDaysLeftPercentage())
